Rename `*_codepage.ipp` to `_converter.hpp`
No need to use a different extension why may confuse tooling.
Also the files contain converter classes, so the naming is confusing.
